requirements.txt
pandas
numpy
scikit-learn
python-dotenv
pymongo==4.6.1
certifi
# pymongo[srv]==3.6


# -e .
